Modified formatting, added equation numbers and added figures
Modified the formatting to include the chapter number in the figures, which necessitated modofication of the heading styles.
Added equation numbers with the chapter number preox.
Added several figures to the current coupler chapter.
Wrote several paragraphs in the current coupler chapter - including the basic derivation for the voltage and current distribution.